scheme-advisor/requirements.txt
Flask==3.0.0
Pillow==10.2.0
groq==0.4.2
python-dotenv==1.0.0
moviepy==1.0.3
gtts==2.5.1
numpy==1.26.3
requests==2.31.0